+++ js/util.js
@@ -1,37 +1,16 @@
-import {NAMES, COMMENTS, getRandomInteger, getRandom} from './data.js';
+import {DESCRIPTION, randomId, randomUrl, randomMessage, getRandomArrayElement, getRandomInteger} from './data.js';
 
-const randomId = getRandom(1, 25);
-const randomUrl = getRandom(1, 25);
-const randomCommentId = getRandom(1, 10000000);
+const descriptionPhoto = () => ({
+  id: randomId(),
+  url: `photos/${randomUrl()}.jpg`,
+  description: getRandomArrayElement(DESCRIPTION),
+  likes: getRandomInteger(15, 200),
+  сomments: randomMessage(),
 
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+});
 
-function getRandomMessage () {
-  const quantity = getRandomInteger(1, 2);
-  const comment = [];
-  if (quantity === 1) {
-    comment[0] = {
-      id: randomCommentId(),
-      avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-      message: getRandomArrayElement(COMMENTS),
-      name: getRandomArrayElement(NAMES)
-    };
-    return comment;
-  }else {
-    comment[0] = {
-      id: randomCommentId(),
-      avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-      message: getRandomArrayElement(COMMENTS),
-      name: getRandomArrayElement(NAMES)
-    };
-    comment[1] = {
-      id: randomCommentId(),
-      avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-      message: getRandomArrayElement(COMMENTS),
-      name: getRandomArrayElement(NAMES)
-    };
-    return comment;
-  }
-}
+const photo = Array.from({length:25}, descriptionPhoto);
 
-export {getRandomInteger,randomId, randomUrl, getRandomArrayElement, getRandomMessage};
+console.log(photo);
+
+export {descriptionPhoto};
